Add timeline tests for text replacement uniqueness (#1008)
Adding this as a proposal to make timeline text replacement easier to understand. Right now, timeline translations apply in key order, and keys are regexes. This is definitely confusing, and @Akurosia resorted a lot of them in #986 and #987, which creates some risk that translations that used to work no longer work (if they ever did?)

This pull request adds the stipulation that if you have two translations of the same text and both of them match the text, then they must be able to be applied in either order and get the same result.

For example, it's ok to have the text "Magnetism/Repel" and then have a translation for "Magnetism" and "Repel", because you can apply those in either order. It's not ok to have text "Savage Wave Cannon" and then have a translation for "Savage Wave Cannon" and "Wave Cannon" (pre-translation collision). It's also not ok to have the text "Time Eruption", translate "Time Eruption" as "Zeiteruption" and also have a translation for "Eruption" on its own (post-translation collision).

This is a fairly loose way of catching these errors, but it still allows for partial translations where possible. Other options that are more understandable would be only allowing full text translations or full sync translations (between colons), which leads to some small explosion of translations in cases where there's like "(In)" and "(Out)" on various things. However, it is much more straightforward to explain as well.

The other context for this discussion is that I would like to eventually have the translation block handle triggers at runtime too and remove all of the regexDe etc style regexes everywhere. The things that need to get replaced in regexes are targets, effects, and messages (more or less). This makes me wonder if I should break up the translations into more categories to reduce collisions? We could have timeline text, targets, effects, syncs/messages (with the idea that the current timeline syncs would be included across the targets and syncs categories).

The followup to this test is to add a test that says "if a translation locale applies, then it must translate everything" and it would also be easy to warn or error about translations that don't match anything.