raidboss: do not allow ^ or $ in replaceText/Sync
These replacements are likely going to be used to replace the trigger
regexes, and so having ^ and $ will not work in a different context.

This also updates the ecma version to 2018 which supports negative
lookbehind regex.  Previous negative lookbehinds have all been ok
because they've been strings that get converted to regex at runtime.