util: Add an initial find missing translations script
This will eventually be added to point out missing translations
for trigger regex and for timeline syncs and texts as well.
This would be a single place to go look for where there is
missing text.

This script is meant to replace check_translation.py, eventually.

Example usage, to find missing korean translations:
  python util/find_missing_translations.py -l ko

This will print out lines similar to:
  .\ui\raidboss\raidemulator.js:16 "this.sealRegexes = {"
  .\ui\raidboss\raidemulator.js:23 "this.unsealRegexes = {"
  .\ui\raidboss\data\02-arr\raid\t1.js:43 "alertText: {"
  .\ui\raidboss\data\02-arr\raid\t1.js:60 "alertText: {"
  .\ui\raidboss\data\02-arr\raid\t1.js:78 "infoText: {"
  .\ui\raidboss\data\02-arr\raid\t1.js:87 "infoText: {"
  .\ui\raidboss\data\02-arr\raid\t1.js:102 "infoText: {"
  .\ui\raidboss\data\02-arr\raid\t10.js:25 "alertText: {"