raidboss: update find_missing_translations to print regexes
Also, because of how much overlap there was with the
check_regex_translation script, that was folded in here.

It says "missing" if there's not a timelineReplace sync or effect
that applies to the regex.  It says "incorrect" if there's
a regexDe (for example) and the timelineReplace applied to regex
creates a different regex.