oopsy: properly handle zoneRegex (#1060)
Without this change, any zoneRegex that was a translatable object would
cause all regex for that file to be handled.  This meant that anybody
who said oops anywhere in the world (or had oops in their name) would
cause that to show up in the log.  This was...not intended.

OOPS