Common replacement many added all lang
I have this to submit to you. I worked on it locally, and I managed to integrate everything. I noticed some lines integrated or not in some (js), and I don't understand why. I thought it would be easier to have them here. To do this, I need to delete these lines in about 27 files. I hope this will be useful to you. If you agree I would make these requests.

Affected files:
- arr: ifrit nm
- Hw : fractal continuum, the vaults
- Sb: eureka hydatos, royal city rabanastre, yojimbo, seiryu ex, o10n, o11n,
-Shb: amaurot, akademia anyder, holminster, twinning, the copied factory, innocence nm and ex, titania ex, hades ex,    
          e1n, e2n, e3n, e4n,e1s, e2s, e3s, e4s, e8s.