raidboss: fix trigger asynchronicity issues
Pull #1366 refactored the way triggers are run and broke the requirement
that all triggers are synchronous unless there is a delaySeconds or a
promise field in the trigger.

Without this, two triggers on the same line will be ordered like this:
  trigger1 preRun
  trigger2 preRun
  trigger1 alert
  trigger2 alert

Personally, it is just too hard to think about asynchronicity in any
case other than the two explicit cases and most triggers are written
to assume that triggers earlier in file order will have executed
completely before triggers later run.