raidboss: add `timelineNeedsFixing` flag
This flag indicates that the timeline is not in great shape and was
constructed by hand (most likely).  Because of this, it should not
be used for translations.  Having this flag will hide missing
timeline translations in `util/find_missing_translations.py`.

Ideally, a new timeline should be made with `util/make_timeline.py`
and the flag removed.