jobs: clean up gcd initialization
A lot of code in jobs attempts to manually set a default gcd for various
boxes and procs before the real gcd is set.  This leads to duplicated
code and confusing code.  Clean all this up by having gcdSpell/gcdSkill
return a reasonable value at all times.

This is some cleanup that I noticed in #1572 to make it easier for
future changes to this file.