raidboss: fix TEA nisi bugs
This was broken when switching to effectIds.  effectIds need to be
capitalized from the log (whoops).

Also, this moves around the set of ids into variables so that they're
not duplicated.  I also double checked with a script that these all
match the correct effects, have the right ids, and mapped to the right
number.

Fixes #1587.