oopsy: use netRegex for all effects
This is a very indirect patch which is to get rid of non-jobs usages of
gLang.kEffect, so I can use netRegex for effects in jobs, remove
gLang.kEffect entirely, in order to track atonement stacks for #725.

This also contributes to several other ongoing nebulous efforts like
using NetRegexes for oopsy instead of events (#1487) and also cleaning
up the unfortunate gLang data structures (#1221).