fisher: fix Korean fish names
There was an off-by-one error (47 instead of 46) for fish and so it was
returning ore instead of fish names.  OOPS

However, I noticed on further inspection that it doesn't return all the
fish.  I think ItemUICategory is a better filter than ItemSearchCategory
(which maybe includes only marketboard items).  Switch everything over
to ItemUICategory (where ironically the original 47 id is correct).