fix: match more than 2 English letters
Because in Chinese or Korean server, players have their first name only, no last name.
This box can only hold 2 CJK characters or 3 letters, or 1 CJK character and 2 letters.
So we can simply check if there are more than 3 English letters in the first 3 characters.
If true, whether the rest one is or not CJK character, the box would contain.