jobs: show cheese dungeons, and only show in content
This changes well-fed from always showing for < max level characters
which is kind of irritating for blu mages.

This also now shows for dungeons for all characters.  However, since
it only shows out of combat, it should still just be a brief
reminder vs an eternal cheese that haunts you for not using food.