Add giant red unlocked warnings, remove resize_handle files
It turns out that "blue shaded background with test data" is too subtle
for most users.  There are many bug reports of people screenshotting
the test data from an overlay and asking why it's not updated.
Raidboss has instructions (and is the most commonly used), but other
overlays do not.  This adds a GIANT DASHED RED BORDER to any overlay
when unlocked that surely can't possibly look correct.  To further make
the point, it also adds some text in the bottom left corner as well.

Since it'd be nice to translate the text, it's not possible to stick the
html for the "unlocked (please lock!)" text in the html or even in the
resize_handle.js file, as we need to wait until after the language has
been selected.  Also, to avoid having the resizing border affect the
layout, we need to make the body position: absolute.

Because of both of these reasons, the resize logic is a bit more
entangled, and so it's moved into defaults.css and user_config.js.
If a page doesn't include defaults.css, the text will be hidden.