resources: Update Common Job Util Functions
Updates the job to role mapping to instead use the user-defined lists as
a reference when determining job role.
Additionally shortening the function logic for most function calls to
directly check if the user-defined list includes the job.
Adds a quick exit within WSServer code if location is undefined,
primarily for local usecases (such as running tests).
Lastly, adding tests for the logic around job util functions and
creating the jobToRoleMap.