jobs: refactor combo tracker to be a tree, not a map
Dragoon skill combos share some of the same skills, such as
some having Wheeling Thrust -> Fang and Claw and others having
Fang and Claw -> Wheeling Thrust at the end.

The current combo tracker has a global map of all skills by id,
and so both of these abilities get marked as the "final" ability
in the chain, which is incorrect some of the time.

Refactor the combo tracker to have a tree of nodes instead, so that
there is context about the chain of abilities to get there.
This will let #1804 not have to track specific abilities with timers
and will let it use the combo tracker more naturally.