jobs: fix combo tracker calling callback unexpectedly
Pull #1825 forgot an early return if an id was handled in the state
transition.  It meant that valid ids that were both part of a combo and
treated as combo breakers (e.g. Maim) would get handled twice, causing
Maim to end the combo.

The other issue is that AbortCombo reuses StateTransition, and so
the callback should not be called in these cases.

Fixes #1845