raidboss: add implicit trigger overrides (#1891)
Rather than adding an explicit `override` property on triggers,
this change implicitly makes all triggers with the same id override any
previously processed triggers.  For cactbot internal triggers, this
shouldn't be an issue, as all triggers must be unique.  User triggers
will be processed after internal triggers for obvious reasons.

There is a potential present hazard here that users have triggers with
ids that match existing ids, but are not meant to override that trigger,
and will be defaulted into to unexpected behavior.

There is also a future hazard, if ids ever conflict.  Cactbot internal
triggers have a unique prefix per file that will prevent this from
happening internally, but this is still a potential hazard for future
triggers (new cactbot triggers colliding with old user triggers, or
new user triggers colliding with old user triggers).

For these reasons, there's verbose logging when any trigger is
overridden.  We can relax this in the future and hide behind debug
flags, but especially as this is a breaking change it makes sense to be
loud.

Fixes #1786.