raidboss: remove unneeded tts entries (#1942)
tts is automatic.  I think in most cases it's preferable to keep the tts
output the same as the on screen output unless it's dramatically longer
or it's something where it makes sense to have something very short in
tts because the mechanic is urgent.

This leaves unchanged o2 / o3 / o4 which are extremely verbose in their
triggers, which seemed a bit obnoxious to play over tts.