raidboss: convert remaining responses to output strings
Fix up tests to catch this case.  Also, support output strings that are
string values and not just translatable objects in tests.

It's no longer useful to call Responses dynamically from inside a
`response` object, because we can't dynamically know the output strings
that it will need.  This adds some duplication here, unfortunately.
The long term plan is:
https://github.com/quisquous/cactbot/issues/1307#issuecomment-723486775