raidboss: remove group tts
This was originally going to be some way to have tts for discord
purposes, but I think the overlayplugin shared tab does this better.
Additionally, cactbot triggers are designed for individuals and it
seems pretty clear that nobody has time to edit a million triggers
to behave properly in this group tts mode.