config: fix issues where some triggers weren't being shown
It's possible that this never worked for responses, as responses aren't
in the detail keys.  At any rate, the underlying goal is to not show
triggers that don't have output in them if the developer settings are
off and so just use that directly.

This was preventing things like "E10S Deepshadow Nova" from showing
up without the developer options turned on.