raidboss: Add E10S Head Markers (#2209)
I wasn't able to find a correlation between the the shadow and the
player, other than the shadow has the same job as the player spawning
it. We can make do with some callouts for this, but it'd be nice to
have something more robust for groups running multiple of the same job.