Add support for user folders (#2304)
Fixes #860.

This also adds (currently optional) support for passing in overlayName
as a parameter.  This is required to get subdirectory support and can
be removed in the future.  It's just here for now as a backwards compat
way to avoid C# errors with a newer plugin and older js.

If overlayName is passed, then it will filter files by that name and
also search subdirectories.

Co-authored-by: Panic Stevenson <panic.stevenson@gmail.com>