coverage: add oopsy to coverage report
The main "expansion" table is a little confusing because oopsy is tacked
on at the end, as if it's a separate content type.  We could add a
separate oopsy table, but that's overkill for what most people care
about.

I think it would make sense to defer fixing up the table until the
coverage report handles translations as well, and we can see what that
looks like overall.