test: speed up timeline by reducing n in O(n^2) op
The pre-replacement collision tests that two regexes that apply to the
same text can be applied in either order without causing issues.  In
these cases, we only need to consider the regexes that apply to
a particular item, and not every single regex.

This is about a 20% speedup.  Unfortunately, there's no way to make
the post-collision test any better.

Mentioned in #2483.