raidboss: improve ui for triggers that can't be configured
This fixes a previous TODO in an effort to address issues that
were brought up in #2503.

This adjusts CSS because having the warning emoji makes the line
slightly taller and makes the (previously existing, but harder
to spot) misalignment much more obvious.

This also adds translations for things like `(function)` that were
not translated previously (sorry).