add guard aoes and earlier bigAoe warnings
Without supressSeconds > 3, you would have 8 callouts and an overlapping callout with the bigAoE during second set. This reduces the number of callouts to 4 and does not overlap with the timelineTrigger for Empyrean Iniquity. A timelineTrigger would have to be more complex to account for the same spell names in the timeline during Queen's Guard and The Queen encounters.

I chose to add Multiple to try and make it clear that as one goes off another should follow shortly. Guard's AOEs may not have be clear enough as two Queen's Guards do cast at a time for the first two sets. String is rather long or may produce long tts as is.