raidboss: e12s chiseled sculpture triggers (alternative)
We were running a bunch of e12s and so I had time to test this out
and iron it out.

I personally found that having a trigger on:
* initial number
* left right
* laser bait
* other three lasers
* knockback location

...was just incredibly noisy.

I have cut this back from #2443 to say:
* initial number (inner outer)
* left/right OR left/right + bait #1
* bait #n laser (no other lasers called out)
* knockback location

Inner/Outer seems to work with the changes that I made here to collect
the sculptures separately from the tether.

(@valarnin, sorry to implement a patch that you were in the middle
of, but I wanted to get this in before I didn't have a chance to
test it anymore.)

Closes #2443.