typescript: remove tsconfig noImplicitReturns
When functions have return types, TypeScript already handles the case
where a function cannot have an undefined return type, implicit or
explicit.

I think it noisy to have to add a `return;` at the end of a function
that can return undefined as a type.  This will especially be the
case for triggers, which use implicit return undefined as a way
to not output things in many trigger functions.

Turning this off for now will help make that conversion easier.
If this creates an issue, then once everything has turned into
TypeScript we can revisit and turn it back on.