Fix recursion error from overlay_plugin_api.ts
After #2611, the raid emulator enters infinite recursion.
```
overlay_plugin_api.ts:97 Uncaught (in promise) RangeError: Maximum call stack size exceeded
    at Object.addOverlayListener [as addOverlayListenerOverride] (overlay_plugin_api.ts:97)
    at Object.addOverlayListener [as addOverlayListenerOverride] (overlay_plugin_api.ts:98)
    at Object.addOverlayListener [as addOverlayListenerOverride] (overlay_plugin_api.ts:98)
    at Object.addOverlayListener [as addOverlayListenerOverride] (overlay_plugin_api.ts:98)
```

This is because of this call sequence:
* original addOverlayListener -> checks for override
* calls RaidEmulatorOverlayApiHook version -> calls original function
* original function checks for override again

The root of the issue is that the "original" function needs to not
include the override.  Fix this by returning an internal function
that does not include the override from the setOverride function.

Also, to avoid having to write this for the add/remove/dispatch,
remove those overrides as they are not used.