util: process and generate triggers branch
As trigger files are diverging more and more from things that users
can copy and paste into javascript file, create a triggers branch
that can be viewed online that has the most recent version of all
of the triggers.

This will allow us to start converting trigger files and still
have a place that users can look to see a JavaScript transpiled
version that they can more easily copy and paste as needed.