util: fix find_missing_translations to report timelineReplace mismatches
This used to work, but when we switched over to netregexes from regexes
this broke and nobody ever noticed.  Unfortunately, there's a lot of
regexes that don't need to be translated, so this adds an
*EXTREMELY* hacky method to detect them.

NetRegexes know what fields that they have on construction, and only
particular fields need translation.  Therefore, if a global flag is
flipped and a NetRegex needs that field, insert a noop character
sequence at the beginning of the NetRegex that we can then detect
in order to know if this NetRegex had such fields when it was
constructed.