util: convert gen_coverage_report.js to emit ts
This fixes a bug from #2531 which removed the output extension from
the generated util/coverage/coverage_report.js file.  This caused
a file without an extension to be generated next to the checked in
empty file.  Thankfully (?!) importing extensionless imported this
bare file first before the js file, and so the coverage report
hasn't been broken.