resources: generate more resources as const (#2795)
By letting simple arrays and objects have an implicit type rather
than defining the type, it lets TypeScript know for sure whether
or not an entry is undefined.  For example, with this change,
ZoneId.SomeZone will always be known to exist.

This fixes the root cause of the revert in #2794.