ui: Stop using global Options variables everywhere
This is a partial version of #2857.  Not dealing with globals just
seems like a good refactoring to have.

The lint is changed for rdmty because changing a file that is ignored
is a warning and we have a maximum number of warnings of zero.
So, instead eslint-disable is added to that file.