raidboss: convert function+this usage in trigger files
These are the three places that are assigning a function() to
data, and then using `this` inside of that function.

In preparation for auto-converting triggers to use arrow functions,
and an eslint rule, convert these to arrow functions.

In the future, this form of function+this will still be ok even with
the eslint rule, but it makes the auto-conversions fully automatic.

I also think it's cleaner.