Jobs: Flash proc box when expired and in combat
In case a proc box expires (reaches a duration of 0) and we are still
in combat, continuously flash the box to draw attention to it.

This is intended for proc boxes like dots or self-bufs which should be
always active/running while in combat.

The feature can be turned on in the Jobs settings via the variable
FlashExpiredProcsInCombat.

Proc boxes that should flash should pass
  flashWhenExpired: true
to the addProcBox call that creates them.