Jobs: Add notification sound to proc boxes
In case a visual animation when a proc box expired isn't enough or
desired, playing a short alert sound can be helpful.

By default sounds are disabled.

If sounds are desired, they can be configured to be played either when
the threshold of an ability is reached or if the ability expired.