Jobs: Remove obsolete timerbox reset calls
In the past it seems it was necessary to set the duration of a
timerbox to 0 before setting the new duration. Most likely in order to
somehow reset the timerbox.

As setting a new duration will always result in calling the reset
function of a timerbox, this shouldn't be necessary anymore.

Removing the unnecessary calls will also prevent that the onReset
callbacks will be called more often than necessary.