config: add link to coverage report (#3146)
Requested by @MyTechnoHunter because this link is hard to find.

Also add support for html "inputs".  Maybe this is a case where we need
some other way to add more text at the beginning of a section, but we
can figure that out if we get more examples.

I left this as the raw link text to make it super clear that it was
a link to click on, and what that link was so you could find it outside
of ACT.

It would be really nice if OverlayPlugin supported something like the
`cactbotChooseDirectory` overlay handler for opening links in a browser
outside of ACT.  I would use it for this and view source if it existed.