util: support FIXME to mark improvable translations
Something that comes up every once in a while is that it can be useful
to update the text of a trigger to be slightly better or slightly
different.  The translations for this trigger are not necessarily wrong
(and using what's there is better than English), but it's the kind of
thing where you want to mark it for somebody to come back and look at
later without deleting it.

This patch supports adding `// FIXME` commments on the lines of
translations you want to leave but mark as needing re-translation.