oopsy: replace damageRegex with netRegex
The difference between damageRegex and abilityRegex is that damageRegex
verifies that a particular ability usage results in flag fields that
look like damage.  So, to replace this we use some `playerDamageFields`
that can be added to `NetRegexes.abilityFull`.

Subtle behavior change note: while we're here, this also checks to make
sure that the target is a player, since we (generally) only care about
mistakes where damage hits a player.  Usually cases where we care about
abilities going off (like interrupts) are done with `abilityRegex`.
However, keep a close eye that all converted triggers are of this type.

Possibly in the future we could also make `playerDamageFields` check
that damage > 0, but this seemed like even more change than was needed.

This PR should remove the last uses of oopsy events outside of oopsy
code itself, and ideally in the future we can just remove them.