raidboss: Enforce Non-TriggerOutput Return Defined (#3278)
For all non-(Partial)TriggerOutputs, ensure the return value is defined
if the field is defined. The exception being conditions, potentially
only for the time being if we decide that all trigger data booleans
should be defined within initData.

This only applies to built in triggers, as popup text still handles the
undefined case for user triggers and overrides.