resources: invert netlog_defs fields, change keys
The goal here is reduce the 4 copies of network log definitions
into a single one.  I think netlog_defs has the most information
and so probably should be the source of truth.

net_fields.d.ts uses name: number for typing, and so something like
`typeof logDefinitions.networkStartsCasting.fields` can now replace it.
Similarly, LineEvent0x*.ts uses the same name: number style and can
be replaced by this.

netregexes.ts and log splitter will need data to be massaged, so the
format doesn't matter as much, so long as it is in a common place.

Followups here will be to switch net_fields, LineEvent, and netregexes
to use this same source of information, and to rename fields and lines
as necessary to make it all play nice together.