docs: Fix issues with hashes on ACT log line examples (#3425)
Fixes #3405

Might be merge conflicts between this PR and #3423, I'm not sure. If there are, I'd rather this PR get rebased/merged second because it's the simpler of the two and will be easier to rebase.

Also includes the other adjustment mentioned in the linked issue to make things more uniform in LineEvent classes.

Also also fixes 0x00 and 0x0C lines having their last structure parameter dropped.