raidboss: remove use of CaresAbout from triggers (#3432)
There have been numerous requests and confusion over time around this
issue.  In particular, it's not clear to most people why sometimes
triggers show up and sometimes they don't (causing them to file bugs or
ask in discord).

This was added in the long ago past when there was no config UI and it
was harder to disable triggers.  Now that it's easy to just click off
anything somebody dislikes, avoiding being too noisy is much less of
a goal.

I considered the alternative of adding an option for this, but I think
there are already WAY too many options in cactbot and I think this would
be confusing.  I'd rather just funnel everybody towards the easier to
explain path of "if you don't like it, turn it off in the UI".