plugin: Change ACT plugin ToProperCase to C# ToTitleCase (#3513)
This function is going away in a future version of the FFXIV
ACT plugin, and so remove it from cactbot now.  The plugin
version also handles capitalization of roman numerals.
Since we are only using this for player names, ignore this.

See:
https://discord.com/channels/551474815727304704/816017800723955792/894766881888669767

It's not clear to me why we need to capitalize these at all.
It was added in c73a19a37e638ec542bc34c1e344c374c044e333,
but without any comments.