raidboss: remove manual role conditions (#3534)
This is a follow-up to #3432.  That PR removed all of the "cares about"
condition checks.  This PR removes all of the manual data.role checks.

Things that are healing only (e.g. "heal to full" or "heal this person")
are left as healer only.

Shared tankbusters (for now) are left as not calling out for dps to
avoid anybody seeing the word "shared" and joining in.  I would like to
follow up with a `Response.sharedTankBuster` that tries to handle this
a little bit more elegantly.