raidboss: add config/testing early out for e12s lion tethers
The cactbot config UI calls all response functions in order to get
the output strings to display.  However, it doesn't have any state
at all, and so all response functions should be able to be called
with nothing in `data` and have no errors or output.

This prevents "Unable to locate a valid lion" from appearing when
starting cactbot. This is a fixup for #3571.