raidboss: improve alerts for "Formation: Sharp Turn" (#3583)
During Operated Superior Flight Units fight of The Puppets Bunker, the
units use a formation attack called "Formation: Sharp Turn" This attack
consists of the 3 units moving into a triangle formation, and then
charging across the screen. The first part of the attack consists of
a single setup skill, 4FAB which cactbot already has a trigger for.
After this resolves, the units then use a separate set of skills that
will be either a left-side based strike or a right-side based strike
charge.

Add triggers for these two abilities to allow alerting where the safe
spot is. For the 5513/5511/4FA9 set of abilities, the safe spot will be
to stand in the center of the arena between all 3 units. For the
5514/5512/4FAA abilities, the safe spot is to stand behind the units.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>