oopsy: add death reports (#3579)
This is a rough initial version.  It replaces the previous "last damage"
tracking and instead tracks a window of events per player.  The events
that it tracks right now are damage, heals, simple mistakes (damageFail
and damageWarn oopsy sets), and missed buffs and mitigation.

Obvious followups here are:
* track death reasons, and other oopsy mistakes
* incorporate hp level into the table of output (somewhere)
* list which effects were active during damage (somehow)
* collapse and trim the report lines to be more readable
* copy and paste of death reports (in markdown, for discord?)

Oopsy shows mistakes for players outside of your party (e.g. in alliance
raids, or DRS as a worst case scenario), and so in order to continue
showing deaths it also tracks events for these players as well.
Hopefully this won't result in a performance issue?  Worst case scenario
would be that you only get death reports for your party and we go back
to special-casing "remember last damage" for non-party members.

The summary page UI is that there's a collapsable chevron, and clicking
on the death line anywhere will expand/collapse the death report
section.

The live list UI is that when people die, their deaths will appear for N
seconds (configurable).  If multiple people die, their deaths will queue
up.  Since we already have behavior for clicking on the live list (which
copies the text), the new behavior is that mousing over the live list
will show that death report.  If in combat, it will fade eventually,
but if out of combat it will appear permanently until the close button
is clicked.