raidboss: add triggers for Meng-Zi/Xun-Zi Deploy Armaments
During the Xun-Zi/Meng-Zi robot fight of the Tower at Paradigm's Rift,
the two robots use some line aoe attacks. These attacks, "Deploy
Armaments" can either be a single big line down the middle of the arena,
or two lines down the sides of the arena.

Add support to trigger an alert on where the safe area is for these
attacks. Use the heading of the skills to determine whether the middle
line attack should be avoided by going north/south or east/west.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>