simplify handling by using a javascript object
Use a new DeployArmaments object to capture the state for Deploy
Armaments. Hold simple boleans indicating if this is a sides attack and
whether ir not the line is vertical. Also store the parsed "finish" time
of the ability.

Modify the collector trigger to generate objects instead of simple
enumerated values. Store the timestamp as well as the two booleans
indicating the attack type.

Modify the alert trigger to use a simpler if-then-else lookup. Reduce
the scope to indicate only the 4 "squares" near the two boss spawns:

*) Go North - Head to the north side of arena
*) Go West - Head to western side of arena
*) Go to North Boss - Stand in front of north boss
*) Go to West Boss - Stand in front of the west boss
*) Go to center - Stand in the center square
*) Go to corner - Stand in the corner between the two bosses.

To allow for better messaging when the mechanics overlap, keep track of
"active" attacks until they finish casting. Use a 3rd cleanup trigger
which checks for and removes all abilities which are no longer finishing
in the future. This makes the trigger say things like:

Go North
(wait 3 seconds)
Go to Corner

When there is an overlapping mechanic where its a middle line aoe from
the east and then another middle line aoe from the north.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>