raidboss: add triggers for Meng-Zi/Xun-Zi Deploy Armaments (#3595)
* raidboss: add triggers for Meng-Zi/Xun-Zi Deploy Armaments

During the Xun-Zi/Meng-Zi robot fight of the Tower at Paradigm's Rift,
the two robots use some line aoe attacks. These attacks, "Deploy
Armaments" can either be a single big line down the middle of the arena,
or two lines down the sides of the arena.

Add support to trigger an alert on where the safe area is for these
attacks. Use the heading of the skills to determine whether the middle
line attack should be avoided by going north/south or east/west.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* use a collector/trigger setup to handle overlapping AOEs

Instead of using separate triggers for each ability, use a single
collector that grabs every AOE heading and type into one of the
4 categories. Then use a delayed trigger to process these into the
relevant message to inform the player.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* simplify handling by using a javascript object

Use a new DeployArmaments object to capture the state for Deploy
Armaments. Hold simple boleans indicating if this is a sides attack and
whether ir not the line is vertical. Also store the parsed "finish" time
of the ability.

Modify the collector trigger to generate objects instead of simple
enumerated values. Store the timestamp as well as the two booleans
indicating the attack type.

Modify the alert trigger to use a simpler if-then-else lookup. Reduce
the scope to indicate only the 4 "squares" near the two boss spawns:

*) Go North - Head to the north side of arena
*) Go West - Head to western side of arena
*) Go to North Boss - Stand in front of north boss
*) Go to West Boss - Stand in front of the west boss
*) Go to center - Stand in the center square
*) Go to corner - Stand in the corner between the two bosses.

To allow for better messaging when the mechanics overlap, keep track of
"active" attacks until they finish casting. Use a 3rd cleanup trigger
which checks for and removes all abilities which are no longer finishing
in the future. This makes the trigger say things like:

Go North
(wait 3 seconds)
Go to Corner

When there is an overlapping mechanic where its a middle line aoe from
the east and then another middle line aoe from the north.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* remove need for third cleanup trigger

Clean up deployArmaments attacks directly within the alert trigger. This
is correct because it doesn't matter when they get removed as long as
its before we actually alert. We know that a new start-cast can ignore
any triggers which will have finished by the time the cast starts.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>