raidboss: add trigger for Bloody Sweep
Hansel and Gretel use "Tandem Assault: Bloody Sweep" as a combo move
which will either sweep the outsides and mainly the south half of the
arena, or sweep the insides hitting mainly the north half of the arena.
Players need to move south or north depending on the position of Hansel
and Gretel. Sometimes, the two bosses will use "Transference" and switch
places.

Add a trigger which detects the relevant IDs and issues an alert on
whether ot move south or north. We can be clever by using a few facts:

* Hansel is left handed, and Gretel is right handed
* The ID indicates which position Hansel or Gretel will be at start
* The cast time can determine whether or not they switch places

Using this information we can use a single strigger which determines the
side of the arena which is safe.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>