raidboss: change casttime check for Bloody Sweep
The cast time for Bloody Sweep is a simple way to check whether it will
be a cast with Transference or not. Use 10 instead of 12, since its the
average between the 7.7 short cast time and the 12.7 long cast time.
This is a bit more clear and might help in the event that cast times are
slightly adjusted in the future.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>