raidboss: add trigger for Bloody Sweep (#3605)
* raidboss: add trigger for Bloody Sweep

Hansel and Gretel use "Tandem Assault: Bloody Sweep" as a combo move
which will either sweep the outsides and mainly the south half of the
arena, or sweep the insides hitting mainly the north half of the arena.
Players need to move south or north depending on the position of Hansel
and Gretel. Sometimes, the two bosses will use "Transference" and switch
places.

Add a trigger which detects the relevant IDs and issues an alert on
whether ot move south or north. We can be clever by using a few facts:

* Hansel is left handed, and Gretel is right handed
* The ID indicates which position Hansel or Gretel will be at start
* The cast time can determine whether or not they switch places

Using this information we can use a single strigger which determines the
side of the arena which is safe.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* raidboss: change casttime check for Bloody Sweep

The cast time for Bloody Sweep is a simple way to check whether it will
be a cast with Transference or not. Use 10 instead of 12, since its the
average between the 7.7 short cast time and the 12.7 long cast time.
This is a bit more clear and might help in the event that cast times are
slightly adjusted in the future.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* raidboss: use "North" and "South" for Bloody Sweep

Instead of using separate "Go North" and "Go South" outputs, just use
the standard "North" and "South" outputs for simplicity.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* remove the unnecessary delay adjustment for Bloody Sweep

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* comment the rationale for Bloody Sweep timing and IDs

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>