oopsy: add current hp to death reports (#3608)
This shows the current hp in the death report on each line prior
to taking damage.  Because current hp is polled from memory, this
value can be incorrect if the player takes a bunch of damage in
a row and so duplicate hp values try to estimate the correct value
by applying damage amounts.  This seems to be ~ok in practice,
but is a bit sketchy.